+++ src/main/java/TRaMis8khae/starbucks/product/domain/ProductOptionCategory.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class ProductOptionCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long productCategoryId;
+    private Integer productCategoryId;
     private Long subCategoryId;
     private Long mainCategoryId;
